Prevent the API from being require()d twice
Having more than one instance of ecnet running at the same time is
problematic. Hack a fix for this by checking if the "ecnet" package has
been loaded and return if instead and setting the "ecnet" global package
to the current program if not (regardless of the actual require string).
This also fixes multiple instances from dofile() since trying to use it
will result in an error.